move all booking pages to use-page-wrapper route group
