+++ apps/web/app/future/org/[orgSlug]/[user]/embed/page.tsx
@@ -1,6 +1,6 @@
 import withEmbedSsrAppDir from "app/WithEmbedSSR";
 import { WithLayout } from "app/layoutHOC";
-import type { OrgPageProps } from "app/org/[orgSlug]/[user]/page";
+import type { OrgPageProps } from "app/(use-page-wrapper)/org/[orgSlug]/[user]/page";
 
 import { getServerSideProps } from "@lib/org/[orgSlug]/[user]/getServerSideProps";
